math

+++ evenOrOdd.py
@@ -1,5 +1,5 @@
-x=int(input("Input a number"))
+x=float(input("Input a number"))
 if x % 2 == 0:
     print("Even")
 else:
-    print('Odd')
+    print("Odd")
